[nxp toup] hostap: fix uAP UDP RX low throughput
The default Max A-MPDU length is 8191 shown in the VHT Cap Info from
uAP's beacon/Assoc response, and this length is read from
hapd->iface->conf->vht_capab. It will lead to the AMPDU only contains
5 packets and the TP is only 60Mbps. Now change the default vht_capab
to VHT_CAP_MAX_A_MPDU_LENGTH_EXPONENT_MAX (1048575), and the TP can
be increased to 83Mbps.

Signed-off-by: Maochen Wang <maochen.wang@nxp.com>