[nxp noup] boards: frdm_rw612: Fix flash read address for NS images
Flash read address was wrong because the flexspi driver and overall
build both are trying to use the 2nd reg entry of the flexspi node for
different purposes. The former expects it to mean the AHB start address
of the flexspi region mapping, the latter expects it to mean the start
address of the text section when XIP. Workaround for now is to delete
the line that overrides it so that the flexspi driver gets what it
wants, and then manually set the CONFIG_FLASH_BASE_ADDRESS to the
location of the NS image partition for NS builds.

Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
Signed-off-by: Waqar Tahir <waqar.tahir@nxp.com>